move select_best_index to vm/src/expr.rs; get rid of OpCmpIdx
