+++ crates/sats/src/relation.rs
@@ -4,11 +4,8 @@ use crate::db::error::RelationError;
 use crate::satn::Satn;
 use crate::{algebraic_type, AlgebraicType, ProductType, ProductTypeElement, Typespace, WithTypespace};
 use derive_more::From;
-use itertools::Itertools;
-use nonempty::NonEmpty;
 use spacetimedb_primitives::{ColId, ColList, ColListBuilder, Constraints, TableId};
 use std::collections::hash_map::DefaultHasher;
-use std::collections::HashSet;
 use std::fmt;
 use std::hash::{Hash, Hasher};
 use std::sync::Arc;
@@ -172,58 +169,6 @@ impl Column {
     }
 }
 
-/// TODO: This is duplicate from `crates/lib/src/operator.rs`
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub enum OpCmpIdx {
-    Eq,
-    NotEq,
-    Lt,
-    LtEq,
-    Gt,
-    GtEq,
-}
-
-impl OpCmpIdx {
-    /// Reverse the order of the `cmp`, to helps in reducing the cases on evaluation, ie:
-    pub fn reverse(self) -> Self {
-        match self {
-            Self::Eq => self,
-            Self::NotEq => self,
-            Self::Lt => Self::Gt,
-            Self::LtEq => Self::GtEq,
-            Self::Gt => Self::Lt,
-            Self::GtEq => Self::LtEq,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct FieldValue<'a> {
-    cmp: OpCmpIdx,
-    field: &'a Column,
-    value: &'a AlgebraicValue,
-}
-
-impl<'a> FieldValue<'a> {
-    pub fn new(cmp: OpCmpIdx, field: &'a Column, value: &'a AlgebraicValue) -> Self {
-        Self { cmp, field, value }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ScanIndex<'a> {
-    Index {
-        cmp: OpCmpIdx,
-        columns: NonEmpty<&'a Column>,
-        value: AlgebraicValue,
-    },
-    Scan {
-        cmp: OpCmpIdx,
-        column: &'a Column,
-        value: AlgebraicValue,
-    },
-}
-
 // TODO(perf): Remove `Clone` impl.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct HeaderOnlyField<'a> {
@@ -359,110 +304,6 @@ impl Header {
                     .any(|(col, ct)| col.contains(find) && ct.contains(&constraint))
             })
             .unwrap_or(false)
-    }
-
-    /// Pick the best index that matches the *permutation* of the supplied `fields`.
-    ///
-    /// This function is designed to handle complex scenarios when selecting the optimal index for a query. The scenarios include:
-    ///
-    /// - Combination of multi and single column indexes that could refer to the same field,
-    ///   so, if we could have `indexes`: `[a], [a, b]` and being asked: `a = 1 AND b = 2 AND a = 3`.
-    /// - Query expression can be supplied in any order, ie: `a = 1 AND b = 2` or `b = 2 AND a = 1`.
-    /// - Different `operators` need to match different `index` operations, ie: `a > 1 AND a = 2 AND a < 3`,
-    ///   must return
-    ///
-    ///   -`ScanIndex::Index(Range(a > 1, a < 3))`
-    ///   -`ScanIndex::Index(a = 2)`
-    /// - The use of multiple tables could generated redundant/duplicate operations like `[ScanIndex::Index(a = 1), ScanIndex::Index(a = 1), ScanIndex::Scan(a = 1)]`.
-    ///   This *can't* be handled here.
-    ///
-    /// # Parameters
-    ///
-    /// - `fields`: A slice of `FieldValue` representing the query conditions.
-    ///
-    /// # Returns
-    ///
-    /// - A vector of `ScanIndex` representing the selected `index` OR `scan` operations.
-    /// - A HashSet of `(FieldName, OpCmpIdx)` representing the fields and operators that where matched by a `index` operation.
-    ///   This is required to remove the redundant operation on `[ScanIndex::Index(a = 1), ScanIndex::Index(a = 1), ScanIndex::Scan(a = 1)]`,
-    ///   that could be generated by calling this function several times by using multiples `JOINS`
-    ///
-    /// # Example
-    ///
-    /// If we have a table with `indexes`: `[a], [b], [b, c]` and then try to
-    /// optimize `WHERE a = 1 AND d > 2 AND c = 2 AND b = 1` we should return
-    ///
-    /// -`ScanIndex::Index(a = 1)`
-    /// -`ScanIndex::Scan(c = 2)`
-    /// -`ScanIndex::Index([c,b] = [1, 2])`
-    ///
-    /// NOTE:
-    ///
-    /// -We don't do extra optimization here (like last the `Scan`), this should be handled in upper layers.
-    /// -We only check up to 4 fields, to keep the permutation small. After that we will miss indexes.
-    pub fn select_best_index<'a>(
-        &'a self,
-        fields: &[FieldValue<'a>],
-    ) -> (Vec<ScanIndex<'a>>, HashSet<(FieldName, OpCmpIdx)>) {
-        let total = std::cmp::min(4, fields.len());
-        let mut found = Vec::with_capacity(total);
-        let mut done = HashSet::with_capacity(total);
-        let mut fields_indexed = HashSet::with_capacity(total);
-
-        let index = Constraints::indexed();
-        for fields in (1..=total).rev().flat_map(|len| fields.iter().permutations(len)) {
-            if fields.iter().any(|x| done.contains(&(x.field, &x.cmp))) {
-                continue;
-            }
-
-            let find = NonEmpty::collect(fields.iter().map(|x| x.field)).unwrap();
-            let find_cols = ColListBuilder::from_iter(fields.iter().map(|x| x.field.col_id))
-                .build()
-                .unwrap();
-            if self
-                .constraints
-                .iter()
-                .any(|(col, ct)| col == &find_cols && ct.contains(&index))
-            {
-                if fields.len() == 1 {
-                    done.extend(fields.iter().map(|x| (x.field, &x.cmp)));
-                    fields_indexed.extend(fields.iter().map(|x| (x.field, &x.cmp)));
-
-                    found.push(ScanIndex::Index {
-                        cmp: fields[0].cmp,
-                        columns: find,
-                        value: fields[0].value.clone(),
-                    });
-                } else if fields.iter().all(|x| x.cmp == x.cmp.reverse()) {
-                    done.extend(fields.iter().map(|x| (x.field, &x.cmp)));
-                    fields_indexed.extend(fields.iter().map(|x| (x.field, &x.cmp)));
-
-                    found.push(ScanIndex::Index {
-                        cmp: fields[0].cmp,
-                        columns: find,
-                        value: ProductValue::from_iter(fields.into_iter().map(|x| x.value.clone())).into(),
-                    });
-                }
-            } else if fields.len() == 1 {
-                let field = fields[0];
-                if !done.contains(&(field.field, &field.cmp)) {
-                    done.insert((field.field, &field.cmp));
-
-                    found.push(ScanIndex::Scan {
-                        cmp: field.cmp,
-                        column: find.head,
-                        value: field.value.clone(),
-                    });
-                }
-            }
-        }
-
-        let done: HashSet<_> = fields_indexed
-            .iter()
-            .map(|(col, cmp)| (col.field.clone(), **cmp))
-            .collect();
-
-        (found, done)
     }
 
     /// Project the [FieldExpr] & the [Constraints] that referenced them
@@ -644,10 +485,8 @@ impl Relation for DbTable {
 
 #[cfg(test)]
 mod tests {
-    use spacetimedb_primitives::col_list;
-
     use super::*;
-    use crate::product;
+    use spacetimedb_primitives::col_list;
 
     /// Build a [Header] using the initial `start_pos` as the column position for the [Constraints]
     fn head(table: &str, fields: (&str, &str), start_pos: u32) -> Header {
@@ -727,353 +566,5 @@ mod tests {
             .unwrap();
 
         assert_eq!(head_rhs, rhs);
-    }
-
-    #[test]
-    fn best_index() {
-        let fields: Vec<_> = ["a", "b", "c", "d", "e"]
-            .iter()
-            .enumerate()
-            .map(|(pos, x)| Column::new(FieldName::named("t1", x), AlgebraicType::I8, pos.into()))
-            .collect();
-
-        let a = ColId(0);
-        let b = ColId(1);
-        let c = ColId(2);
-        let d = ColId(3);
-
-        let col_a = fields[0].clone();
-        let col_b = fields[1].clone();
-        let col_c = fields[2].clone();
-        let col_d = fields[3].clone();
-        let col_e = fields[4].clone();
-
-        let head1 = Header::new(
-            "t1".into(),
-            fields,
-            vec![
-                //Index a
-                (a.into(), Constraints::primary_key()),
-                //Index b
-                (b.into(), Constraints::indexed()),
-                //Index b + c
-                (col_list![b, c], Constraints::unique()),
-                //Index a+ b + c + d
-                (col_list![a, b, c, d], Constraints::indexed()),
-            ],
-        );
-
-        let val_a = AlgebraicValue::U64(1);
-        let val_b = AlgebraicValue::U64(2);
-        let val_c = AlgebraicValue::U64(3);
-        let val_d = AlgebraicValue::U64(4);
-        let val_e = AlgebraicValue::U64(5);
-
-        // Check for simple scan
-        assert_eq!(
-            head1
-                .select_best_index(&[FieldValue::new(OpCmpIdx::Eq, &col_d, &val_e)])
-                .0,
-            vec![ScanIndex::Scan {
-                cmp: OpCmpIdx::Eq,
-                column: &col_d,
-                value: val_e.clone(),
-            }]
-        );
-
-        assert_eq!(
-            head1
-                .select_best_index(&[FieldValue::new(OpCmpIdx::Eq, &col_a, &val_a)])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: NonEmpty::new(&col_a),
-                value: val_a.clone()
-            }]
-        );
-
-        assert_eq!(
-            head1
-                .select_best_index(&[FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b)])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: NonEmpty::new(&col_b),
-                value: val_b.clone()
-            }]
-        );
-
-        // Check for permutation
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c)
-                ])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: (&col_b, vec![&col_c]).into(),
-                value: product![val_b.clone(), val_c.clone()].into()
-            }]
-        );
-
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c),
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b)
-                ])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: (&col_c, vec![&col_b]).into(),
-                value: product![val_c.clone(), val_b.clone()].into()
-            }]
-        );
-
-        // Check for permutation
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c),
-                    FieldValue::new(OpCmpIdx::Eq, &col_d, &val_d)
-                ])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: (&col_a, vec![&col_b, &col_c, &col_d]).into(),
-                value: product![val_a.clone(), val_b.clone(), val_c.clone(), val_d.clone()].into(),
-            }]
-        );
-
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Eq, &col_d, &val_d),
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c)
-                ])
-                .0,
-            vec![ScanIndex::Index {
-                cmp: OpCmpIdx::Eq,
-                columns: (&col_b, vec![&col_a, &col_d, &col_c]).into(),
-                value: product![val_b.clone(), val_a.clone(), val_d.clone(), val_c.clone()].into(),
-            }]
-        );
-
-        // Check mix scan + index
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Eq, &col_e, &val_e),
-                    FieldValue::new(OpCmpIdx::Eq, &col_d, &val_d)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Eq,
-                    columns: NonEmpty::new(&col_b),
-                    value: val_b.clone(),
-                },
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Eq,
-                    columns: NonEmpty::new(&col_a),
-                    value: val_a.clone(),
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Eq,
-                    column: &col_e,
-                    value: val_e.clone(),
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Eq,
-                    column: &col_d,
-                    value: val_d.clone(),
-                }
-            ]
-        );
-
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c),
-                    FieldValue::new(OpCmpIdx::Eq, &col_d, &val_d)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Eq,
-                    columns: (&col_b, vec![&col_c]).into(),
-                    value: product![val_b.clone(), val_c.clone()].into()
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Eq,
-                    column: &col_d,
-                    value: val_d.clone(),
-                }
-            ]
-        );
-    }
-
-    #[test]
-    fn best_index_range() {
-        let fields: Vec<_> = ["a", "b", "c", "d", "e"]
-            .iter()
-            .enumerate()
-            .map(|(pos, x)| Column::new(FieldName::named("t1", x), AlgebraicType::I8, pos.into()))
-            .collect();
-
-        let a = ColId(0);
-        let b = ColId(1);
-        let c = ColId(2);
-        let d = ColId(3);
-
-        let col_a = fields[0].clone();
-        let col_b = fields[1].clone();
-        let col_c = fields[2].clone();
-        let col_d = fields[3].clone();
-
-        let head1 = Header::new(
-            "t1".into(),
-            fields,
-            vec![
-                //Index a
-                (a.into(), Constraints::primary_key()),
-                //Index b
-                (b.into(), Constraints::indexed()),
-                //Index b + c
-                (col_list![b, c], Constraints::unique()),
-                //Index a+ b + c + d
-                (col_list![a, b, c, d], Constraints::indexed()),
-            ],
-        );
-
-        let val_a = AlgebraicValue::U64(1);
-        let val_b = AlgebraicValue::U64(2);
-        let val_c = AlgebraicValue::U64(3);
-        let val_d = AlgebraicValue::U64(4);
-
-        // Same field indexed
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Gt, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Lt, &col_a, &val_b)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Gt,
-                    columns: NonEmpty::new(&col_a),
-                    value: val_a.clone()
-                },
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Lt,
-                    columns: NonEmpty::new(&col_a),
-                    value: val_b.clone()
-                },
-            ]
-        );
-
-        // Same field scan
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Gt, &col_d, &val_d),
-                    FieldValue::new(OpCmpIdx::Lt, &col_d, &val_b)
-                ])
-                .0,
-            vec![
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Gt,
-                    column: &col_d,
-                    value: val_d.clone()
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Lt,
-                    column: &col_d,
-                    value: val_b.clone()
-                }
-            ]
-        );
-        // One indexed other scan
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Gt, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Lt, &col_c, &val_c)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Gt,
-                    columns: NonEmpty::new(&col_b),
-                    value: val_b.clone()
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Lt,
-                    column: &col_c,
-                    value: val_c.clone()
-                }
-            ]
-        );
-
-        // 1 multi-indexed 1 index
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Eq, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::GtEq, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Eq, &col_c, &val_c)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Eq,
-                    columns: (&col_b, vec![&col_c]).into(),
-                    value: product![val_b.clone(), val_c.clone()].into()
-                },
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::GtEq,
-                    columns: NonEmpty::new(&col_a),
-                    value: val_a.clone()
-                },
-            ]
-        );
-
-        // 1 indexed 2 scan
-        assert_eq!(
-            head1
-                .select_best_index(&[
-                    FieldValue::new(OpCmpIdx::Gt, &col_b, &val_b),
-                    FieldValue::new(OpCmpIdx::Eq, &col_a, &val_a),
-                    FieldValue::new(OpCmpIdx::Lt, &col_c, &val_c)
-                ])
-                .0,
-            vec![
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Gt,
-                    columns: NonEmpty::new(&col_b),
-                    value: val_b.clone()
-                },
-                ScanIndex::Index {
-                    cmp: OpCmpIdx::Eq,
-                    columns: NonEmpty::new(&col_a),
-                    value: val_a.clone()
-                },
-                ScanIndex::Scan {
-                    cmp: OpCmpIdx::Lt,
-                    column: &col_c,
-                    value: val_c.clone()
-                }
-            ]
-        );
     }
 }
